+++ Sources/SwiftUICharts/Common/ChartStyle.swift
@@ -1,0 +1,128 @@
+//
+//  File.swift
+//
+//
+//  Created by Pawel on 09/08/2020.
+//
+
+import SwiftUI
+
+public class ChartStyle {
+    public var backgroundColor: Color
+    public var accentColor: Color
+    public var gradientColor: GradientColor
+    public var textColor: Color
+    public var legendTextColor: Color
+    public var dropShadowColor: Color
+    public weak var darkModeStyle: ChartStyle?
+
+    public init(backgroundColor: Color, accentColor: Color, secondGradientColor: Color, textColor: Color, legendTextColor: Color, dropShadowColor: Color) {
+        self.backgroundColor = backgroundColor
+        self.accentColor = accentColor
+        self.gradientColor = GradientColor(start: accentColor, end: secondGradientColor)
+        self.textColor = textColor
+        self.legendTextColor = legendTextColor
+        self.dropShadowColor = dropShadowColor
+    }
+
+    public init(backgroundColor: Color, accentColor: Color, gradientColor: GradientColor, textColor: Color, legendTextColor: Color, dropShadowColor: Color) {
+        self.backgroundColor = backgroundColor
+        self.accentColor = accentColor
+        self.gradientColor = gradientColor
+        self.textColor = textColor
+        self.legendTextColor = legendTextColor
+        self.dropShadowColor = dropShadowColor
+    }
+
+    public init(formSize: CGSize) {
+        self.backgroundColor = Color.white
+        self.accentColor = Colors.OrangeStart
+        self.gradientColor = GradientColors.orange
+        self.legendTextColor = Color.gray
+        self.textColor = Color.black
+        self.dropShadowColor = Color.gray
+    }
+}
+
+public extension ChartStyle {
+    static let lineChartStyleOne = ChartStyle(
+        backgroundColor: Color.white,
+        accentColor: Colors.OrangeStart,
+        secondGradientColor: Colors.OrangeEnd,
+        textColor: Color.black,
+        legendTextColor: Color.gray,
+        dropShadowColor: Color.gray
+    )
+
+    static let barChartStyleOrangeLight = ChartStyle(
+        backgroundColor: Color.white,
+        accentColor: Colors.OrangeStart,
+        secondGradientColor: Colors.OrangeEnd,
+        textColor: Color.black,
+        legendTextColor: Color.gray,
+        dropShadowColor: Color.gray
+    )
+
+    static let barChartStyleOrangeDark = ChartStyle(
+        backgroundColor: Color.black,
+        accentColor: Colors.OrangeStart,
+        secondGradientColor: Colors.OrangeEnd,
+        textColor: Color.white,
+        legendTextColor: Color.gray,
+        dropShadowColor: Color.gray
+    )
+
+    static let barChartStyleNeonBlueLight = ChartStyle(
+        backgroundColor: Color.white,
+        accentColor: Colors.GradientNeonBlue,
+        secondGradientColor: Colors.GradientPurple,
+        textColor: Color.black,
+        legendTextColor: Color.gray,
+        dropShadowColor: Color.gray
+    )
+
+    static let barChartStyleNeonBlueDark = ChartStyle(
+        backgroundColor: Color.black,
+        accentColor: Colors.GradientNeonBlue,
+        secondGradientColor: Colors.GradientPurple,
+        textColor: Color.white,
+        legendTextColor: Color.gray,
+        dropShadowColor: Color.gray
+    )
+
+    static let barChartMidnightGreenDark = ChartStyle(
+        backgroundColor: Color(hexString: "#36534D"), // 3B5147, 313D34
+        accentColor: Color(hexString: "#FFD603"),
+        secondGradientColor: Color(hexString: "#FFCA04"),
+        textColor: Color.white,
+        legendTextColor: Color(hexString: "#D2E5E1"),
+        dropShadowColor: Color.gray
+    )
+
+    static let barChartMidnightGreenLight = ChartStyle(
+        backgroundColor: Color.white,
+        accentColor: Color(hexString: "#84A094"), // 84A094 , 698378
+        secondGradientColor: Color(hexString: "#50675D"),
+        textColor: Color.black,
+        legendTextColor: Color.gray,
+        dropShadowColor: Color.gray
+    )
+
+    static let pieChartStyleOne = ChartStyle(
+        backgroundColor: Color.white,
+        accentColor: Colors.OrangeEnd,
+        secondGradientColor: Colors.OrangeStart,
+        textColor: Color.black,
+        legendTextColor: Color.gray,
+        dropShadowColor: Color.gray
+    )
+
+    static let lineViewDarkMode = ChartStyle(
+        backgroundColor: Color.black,
+        accentColor: Colors.OrangeStart,
+        secondGradientColor: Colors.OrangeEnd,
+        textColor: Color.white,
+        legendTextColor: Color.white,
+        dropShadowColor: Color.gray
+    )
+}
