+++ Sources/SwiftUICharts/PieChart/PieChartCell.swift
@@ -20,14 +20,15 @@ public struct PieChartCell: View {
     @State private var show: Bool = false
     var rect: CGRect
     var radius: CGFloat {
-        return min(rect.width, rect.height) / 2
+        min(rect.width, rect.height) / 2
     }
 
     var startDeg: Double
     var endDeg: Double
     var path: Path {
         var path = Path()
-        path.addArc(center: rect.mid, radius: self.radius, startAngle: Angle(degrees: self.startDeg), endAngle: Angle(degrees: self.endDeg), clockwise: false)
+        path.addArc(center: rect.mid, radius: radius, startAngle: Angle(degrees: startDeg),
+                    endAngle: Angle(degrees: endDeg), clockwise: false)
         path.addLine(to: rect.mid)
         path.closeSubpath()
         return path
@@ -36,13 +37,14 @@ public struct PieChartCell: View {
     var index: Int
     var backgroundColor: Color
     var accentColor: Color
+
     public var body: some View {
         path
             .fill()
-            .foregroundColor(self.accentColor)
-            .overlay(path.stroke(self.backgroundColor, lineWidth: 2))
-            .scaleEffect(self.show ? 1 : 0)
-            .animation(Animation.spring().delay(Double(self.index) * 0.04))
+            .foregroundColor(accentColor)
+            .overlay(path.stroke(backgroundColor, lineWidth: 2))
+            .scaleEffect(show ? 1 : 0)
+            .animation(Animation.spring().delay(Double(index) * 0.04))
             .onAppear {
                 self.show = true
             }
@@ -51,7 +53,7 @@ public struct PieChartCell: View {
 
 extension CGRect {
     var mid: CGPoint {
-        return CGPoint(x: self.midX, y: self.midY)
+        CGPoint(x: midX, y: midY)
     }
 }
