+++ Sources/SwiftUICharts/PieChart/PieChartRow.swift
@@ -9,14 +9,21 @@
 import SwiftUI
 
 public struct PieChartRow: View {
-    var data: [Double]
+    public var data: ChartData
     var backgroundColor: Color
     var accentColor: Color
+    var values: [Double] {
+        data.points.map(\.value)
+    }
+
+    var maxValue: Double {
+        values.reduce(0, +)
+    }
+
     var slices: [PieSlice] {
         var tempSlices: [PieSlice] = []
         var lastEndDeg: Double = 0
-        let maxValue = data.reduce(0, +)
-        for slice in data {
+        for slice in values {
             let normalized: Double = Double(slice) / Double(maxValue)
             let startDeg = lastEndDeg
             let endDeg = lastEndDeg + (normalized * 360)
@@ -27,13 +34,13 @@ public struct PieChartRow: View {
     }
 
     @Binding var showValue: Bool
-    @Binding var currentValue: Double
+    @Binding var currentValue: String
 
-    @State private var currentTouchedIndex = -1 {
+    @State private var currentTouchedIndex: Int? = nil {
         didSet {
             if oldValue != currentTouchedIndex {
-                showValue = currentTouchedIndex != -1
-                currentValue = showValue ? slices[currentTouchedIndex].value : 0
+                showValue = currentTouchedIndex != nil
+                currentValue = showValue ? data.points[currentTouchedIndex!].formattedValue : ""
             }
         }
     }
@@ -42,7 +49,7 @@ public struct PieChartRow: View {
         GeometryReader { geometry in
             ZStack {
                 ForEach(0 ..< self.slices.count) { i in
-                    PieChartCell(rect: geometry.frame(in: .local), startDeg: self.slices[i].startDeg, endDeg: self.slices[i].endDeg, index: i, backgroundColor: self.backgroundColor, accentColor: self.accentColor)
+                    PieChartCell(rect: geometry.frame(in: .local), startDeg: self.slices[i].startDeg, endDeg: self.slices[i].endDeg, index: i, backgroundColor: self.backgroundColor, accentColor: self.data.points[i].color)
                         .scaleEffect(self.currentTouchedIndex == i ? 1.1 : 1)
                         .animation(Animation.spring())
                 }
@@ -53,27 +60,27 @@ public struct PieChartRow: View {
                     let isTouchInPie = isPointInCircle(point: value.location, circleRect: rect)
                     if isTouchInPie {
                         let touchDegree = degree(for: value.location, inCircleRect: rect)
-                        self.currentTouchedIndex = self.slices.firstIndex(where: { $0.startDeg < touchDegree && $0.endDeg > touchDegree }) ?? -1
+                        self.currentTouchedIndex = self.slices.firstIndex { $0.startDeg < touchDegree && $0.endDeg > touchDegree } ?? nil
                     } else {
-                        self.currentTouchedIndex = -1
+                        self.currentTouchedIndex = nil
                     }
                 }
                 .onEnded { value in
-                    self.currentTouchedIndex = -1
-                        })
+                    self.currentTouchedIndex = nil
+                })
         }
     }
 }
 
-#if DEBUG
-struct PieChartRow_Previews: PreviewProvider {
-    static var previews: some View {
-        Group {
-            PieChartRow(data: [8, 23, 54, 32, 12, 37, 7, 23, 43], backgroundColor: Color(red: 252.0 / 255.0, green: 236.0 / 255.0, blue: 234.0 / 255.0), accentColor: Color(red: 225.0 / 255.0, green: 97.0 / 255.0, blue: 76.0 / 255.0), showValue: Binding.constant(false), currentValue: Binding.constant(0))
-                .frame(width: 100, height: 100)
-            PieChartRow(data: [0], backgroundColor: Color(red: 252.0 / 255.0, green: 236.0 / 255.0, blue: 234.0 / 255.0), accentColor: Color(red: 225.0 / 255.0, green: 97.0 / 255.0, blue: 76.0 / 255.0), showValue: Binding.constant(false), currentValue: Binding.constant(0))
-                .frame(width: 100, height: 100)
-        }
-    }
-}
-#endif
+// #if DEBUG
+// struct PieChartRow_Previews: PreviewProvider {
+//    static var previews: some View {
+//        Group {
+//            PieChartRow(data: [8, 23, 54, 32, 12, 37, 7, 23, 43], backgroundColor: Color(red: 252.0 / 255.0, green: 236.0 / 255.0, blue: 234.0 / 255.0), accentColor: Color(red: 225.0 / 255.0, green: 97.0 / 255.0, blue: 76.0 / 255.0), showValue: Binding.constant(false), currentValue: Binding.constant(0))
+//                .frame(width: 100, height: 100)
+//            PieChartRow(data: [0], backgroundColor: Color(red: 252.0 / 255.0, green: 236.0 / 255.0, blue: 234.0 / 255.0), accentColor: Color(red: 225.0 / 255.0, green: 97.0 / 255.0, blue: 76.0 / 255.0), showValue: Binding.constant(false), currentValue: Binding.constant(0))
+//                .frame(width: 100, height: 100)
+//        }
+//    }
+// }
+// #endif
